Update Double - BigDecimal

+++ back-end/src/main/java/rccommerce/dto/OperatorDTO.java
@@ -1,27 +1,28 @@
 package rccommerce.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import rccommerce.entities.Operator;
 
 public class OperatorDTO extends UserDTO {
 
-	@NotNull(message = "Campo requerido")
-	@PositiveOrZero(message = "Comissão deve ter um valor zero ou positivo")
-	private Double commission;
+    @NotNull(message = "Campo requerido")
+    @PositiveOrZero(message = "Comissão deve ter um valor zero ou positivo")
+    private BigDecimal commission;
 
-	public OperatorDTO(Long id, String name, String email, String password, Double commission) {
-		super(id, name, email, password);
-		this.commission = commission;
-	}
+    public OperatorDTO(Long id, String name, String email, String password, BigDecimal commission) {
+        super(id, name, email, password);
+        this.commission = commission;
+    }
 
-	public OperatorDTO(Operator entity) {
-		super(entity);
-		commission = entity.getCommission();
-	}
+    public OperatorDTO(Operator entity) {
+        super(entity);
+        commission = entity.getCommission();
+    }
 
-	public Double getCommission() {
-		return commission;
-	}
-
+    public BigDecimal getCommission() {
+        return commission;
+    }
 }
