+++ back-end/src/main/java/rccommerce/dto/OperatorMinDTO.java
@@ -1,23 +1,24 @@
 package rccommerce.dto;
 
+import java.math.BigDecimal;
+
 import rccommerce.entities.Operator;
 
 public class OperatorMinDTO extends UserMinDTO {
 
-	private Double commission;
+    private BigDecimal commission;
 
-	public OperatorMinDTO(Long id, String name, String email, Double commission) {
-		super(id, name, email);
-		this.commission = commission;
-	}
+    public OperatorMinDTO(Long id, String name, String email, BigDecimal commission) {
+        super(id, name, email);
+        this.commission = commission;
+    }
 
-	public OperatorMinDTO(Operator entity) {
-		super(entity);
-		commission = entity.getCommission();
-	}
+    public OperatorMinDTO(Operator entity) {
+        super(entity);
+        commission = entity.getCommission();
+    }
 
-	public Double getCommission() {
-		return commission;
-	}
-	
+    public BigDecimal getCommission() {
+        return commission;
+    }
 }
