+++ back-end/src/main/java/rccommerce/entities/Operator.java
@@ -1,5 +1,7 @@
 package rccommerce.entities;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import rccommerce.dto.OperatorDTO;
@@ -10,31 +12,31 @@ import rccommerce.services.interfaces.Convertible;
 @Table(name = "tb_operator")
 public class Operator extends User implements Convertible<OperatorDTO, OperatorMinDTO> {
 
-	private Double commission;
+    private BigDecimal commission;
 
-	public Operator() {
-	}
+    public Operator() {
+    }
 
-	public Operator(Long id, String name, String email, String password, Double commission) {
-		super(id, name, email, password);
-		this.commission = commission;
-	}
+    public Operator(Long id, String name, String email, String password, BigDecimal commission) {
+        super(id, name, email, password);
+        this.commission = commission;
+    }
 
-	public Double getCommission() {
-		return commission;
-	}
+    public BigDecimal getCommission() {
+        return commission;
+    }
 
-	public void setCommission(Double commission) {
-		this.commission = commission;
-	}
+    public void setCommission(BigDecimal commission) {
+        this.commission = commission;
+    }
 
-	@Override
-	public OperatorDTO convertDTO() {
-		return new OperatorDTO(this);
-	}
+    @Override
+    public OperatorDTO convertDTO() {
+        return new OperatorDTO(this);
+    }
 
-	@Override
-	public OperatorMinDTO convertMinDTO() {
-		return new OperatorMinDTO(this);
-	}
+    @Override
+    public OperatorMinDTO convertMinDTO() {
+        return new OperatorMinDTO(this);
+    }
 }
